- removed some FindBug warnings

+++ jsettlers.logic.objects/jsettlers/logic/objects/building/ConstructionMarkObject.java
@@ -47,7 +47,7 @@ public class ConstructionMarkObject extends AbstractHexMapObject {
 
 	public void setConstructionValue(byte constructionValue) {
 		assert constructionValue >= 0 : "construction value must be >= 0";
-		this.constructionValue = constructionValue / Byte.MAX_VALUE;
+		this.constructionValue = ((float) constructionValue) / Byte.MAX_VALUE;
 	}
 
 }
